Move `asset-swapper` related options to `options.ts` from `config.ts`
* As `logger.ts` depends on `config.ts`, `config.ts` depending on
`asset-swapper` makes it difficult to avoid circular dependency.
