+++ src/config.ts
@@ -1,4 +1,5 @@
 import { assert } from '@0x/assert';
+import { ChainId } from '@0x/contract-addresses';
 import { TokenMetadatasForChains, valueByChainId } from '@0x/token-metadata';
 import { BigNumber } from '@0x/utils';
 import * as fs from 'fs';
@@ -7,20 +8,8 @@ import { linearBuckets } from 'prom-client';
 import * as validateUUID from 'uuid-validate';
 
 import {
-    BlockParamLiteral,
-    ChainId,
-    DEFAULT_TOKEN_ADJACENCY_GRAPH_BY_CHAIN_ID,
-    OrderPrunerPermittedFeeTypes,
-    SamplerOverrides,
-    SOURCE_FLAGS,
-    SwapQuoteRequestOpts,
-    SwapQuoterOpts,
-    SwapQuoterRfqOpts,
-} from './asset-swapper';
-import {
     DEFAULT_LOGGER_INCLUDE_TIMESTAMP,
     DEFAULT_META_TX_MIN_ALLOWED_SLIPPAGE,
-    DEFAULT_QUOTE_SLIPPAGE_PERCENTAGE,
     DEFAULT_ZERO_EX_GAS_API_URL,
     HEALTHCHECK_PATH,
     METRICS_PATH,
@@ -28,7 +17,6 @@ import {
     ONE_HOUR_MS,
     ONE_MINUTE_MS,
     ORDERBOOK_PATH,
-    QUOTE_ORDER_EXPIRATION_BUFFER_MS,
     TEN_MINUTES_MS,
     TX_BASE_GAS,
 } from './constants';
@@ -263,7 +251,7 @@ export const RFQT_REGISTRY_PASSWORDS: string[] = resolveEnvVar<string[]>(
     [],
 );
 
-const RFQT_INTEGRATORS: Integrator[] = INTEGRATORS_ACL.filter((i) => i.rfqt);
+export const RFQT_INTEGRATORS: Integrator[] = INTEGRATORS_ACL.filter((i) => i.rfqt);
 export const RFQT_INTEGRATOR_IDS: string[] = INTEGRATORS_ACL.filter((i) => i.rfqt).map((i) => i.integratorId);
 export const RFQT_API_KEY_WHITELIST: string[] = getApiKeyWhitelistFromIntegratorsAcl('rfqt');
 
@@ -359,111 +347,13 @@ const UNWRAP_WETH_GAS = UNWRAP_GAS_BY_CHAIN_ID[CHAIN_ID];
 export const UNWRAP_QUOTE_GAS = TX_BASE_GAS.plus(UNWRAP_WETH_GAS);
 export const WRAP_QUOTE_GAS = UNWRAP_QUOTE_GAS;
 
-const FILL_QUOTE_TRANSFORMER_GAS_OVERHEAD = new BigNumber(150e3);
-const EXCHANGE_PROXY_OVERHEAD_NO_VIP = () => FILL_QUOTE_TRANSFORMER_GAS_OVERHEAD;
-const MULTIPLEX_BATCH_FILL_SOURCE_FLAGS =
-    SOURCE_FLAGS.Uniswap_V2 |
-    SOURCE_FLAGS.SushiSwap |
-    SOURCE_FLAGS.RfqOrder |
-    SOURCE_FLAGS.Uniswap_V3 |
-    SOURCE_FLAGS.OtcOrder;
-const MULTIPLEX_MULTIHOP_FILL_SOURCE_FLAGS = SOURCE_FLAGS.Uniswap_V2 | SOURCE_FLAGS.SushiSwap | SOURCE_FLAGS.Uniswap_V3;
-const EXCHANGE_PROXY_OVERHEAD_FULLY_FEATURED = (sourceFlags: bigint) => {
-    if ([SOURCE_FLAGS.Uniswap_V2, SOURCE_FLAGS.SushiSwap].includes(sourceFlags)) {
-        // Uniswap and forks VIP
-        return TX_BASE_GAS;
-    } else if (
-        [
-            SOURCE_FLAGS.SushiSwap,
-            SOURCE_FLAGS.PancakeSwap,
-            SOURCE_FLAGS.PancakeSwap_V2,
-            SOURCE_FLAGS.BakerySwap,
-            SOURCE_FLAGS.ApeSwap,
-        ].includes(sourceFlags) &&
-        CHAIN_ID === ChainId.BSC
-    ) {
-        // PancakeSwap and forks VIP
-        return TX_BASE_GAS;
-    } else if (SOURCE_FLAGS.Uniswap_V3 === sourceFlags) {
-        // Uniswap V3 VIP
-        return TX_BASE_GAS.plus(5e3);
-    } else if (SOURCE_FLAGS.Curve === sourceFlags) {
-        // Curve pseudo-VIP
-        return TX_BASE_GAS.plus(40e3);
-    } else if (SOURCE_FLAGS.RfqOrder === sourceFlags) {
-        // RFQ VIP
-        return TX_BASE_GAS.plus(5e3);
-    } else if (SOURCE_FLAGS.OtcOrder === sourceFlags) {
-        // OtcOrder VIP
-        // NOTE: Should be 15k cheaper after the first tx from txOrigin than RfqOrder
-        // Use 5k less for now as not to over bias
-        return TX_BASE_GAS;
-    } else if ((MULTIPLEX_BATCH_FILL_SOURCE_FLAGS | sourceFlags) === MULTIPLEX_BATCH_FILL_SOURCE_FLAGS) {
-        if ((sourceFlags & SOURCE_FLAGS.OtcOrder) === SOURCE_FLAGS.OtcOrder) {
-            // Multiplex that has OtcOrder
-            return TX_BASE_GAS.plus(10e3);
-        }
-        // Multiplex batch fill
-        return TX_BASE_GAS.plus(15e3);
-    } else if (
-        (MULTIPLEX_MULTIHOP_FILL_SOURCE_FLAGS | sourceFlags) ===
-        (MULTIPLEX_MULTIHOP_FILL_SOURCE_FLAGS | SOURCE_FLAGS.MultiHop)
-    ) {
-        // Multiplex multi-hop fill
-        return TX_BASE_GAS.plus(25e3);
-    } else {
-        return FILL_QUOTE_TRANSFORMER_GAS_OVERHEAD;
-    }
-};
-
-const NEON_ROUTER_NUM_SAMPLES = 14;
-// TODO(kimpers): Due to an issue with the Rust router we want to use equidistant samples when using the Rust router
 const DEFAULT_SAMPLE_DISTRIBUTION_BASE = 1;
-
-const SAMPLE_DISTRIBUTION_BASE: number = _.isEmpty(process.env.SAMPLE_DISTRIBUTION_BASE)
+export const SAMPLE_DISTRIBUTION_BASE: number = _.isEmpty(process.env.SAMPLE_DISTRIBUTION_BASE)
     ? DEFAULT_SAMPLE_DISTRIBUTION_BASE
     : assertEnvVarType('SAMPLE_DISTRIBUTION_BASE', process.env.SAMPLE_DISTRIBUTION_BASE, EnvVarType.Float);
 
-export const ASSET_SWAPPER_MARKET_ORDERS_OPTS: Partial<SwapQuoteRequestOpts> = {
-    bridgeSlippage: DEFAULT_QUOTE_SLIPPAGE_PERCENTAGE,
-    numSamples: 13,
-    sampleDistributionBase: SAMPLE_DISTRIBUTION_BASE,
-    neonRouterNumSamples: NEON_ROUTER_NUM_SAMPLES,
-    exchangeProxyOverhead: EXCHANGE_PROXY_OVERHEAD_FULLY_FEATURED,
-    shouldGenerateQuoteReport: true,
-};
-
-export const ASSET_SWAPPER_MARKET_ORDERS_OPTS_NO_VIP: Partial<SwapQuoteRequestOpts> = {
-    ...ASSET_SWAPPER_MARKET_ORDERS_OPTS,
-    exchangeProxyOverhead: EXCHANGE_PROXY_OVERHEAD_NO_VIP,
-};
-
 export const CHAIN_HAS_VIPS = (chainId: ChainId) => {
     return [ChainId.Mainnet, ChainId.BSC].includes(chainId);
-};
-
-const SAMPLER_OVERRIDES: SamplerOverrides | undefined = (() => {
-    switch (CHAIN_ID) {
-        case ChainId.Ganache:
-            return { overrides: {}, block: BlockParamLiteral.Latest };
-        default:
-            return undefined;
-    }
-})();
-
-const SWAP_QUOTER_RFQT_OPTS: SwapQuoterRfqOpts = {
-    integratorsWhitelist: RFQT_INTEGRATORS,
-    txOriginBlacklist: RFQT_TX_ORIGIN_BLACKLIST,
-};
-
-export const SWAP_QUOTER_OPTS: Partial<SwapQuoterOpts> = {
-    chainId: CHAIN_ID,
-    expiryBufferMs: QUOTE_ORDER_EXPIRATION_BUFFER_MS,
-    rfqt: SWAP_QUOTER_RFQT_OPTS,
-    zeroExGasApiUrl: ZERO_EX_GAS_API_URL,
-    permittedOrderFeeTypes: new Set([OrderPrunerPermittedFeeTypes.NoFees]),
-    samplerOverrides: SAMPLER_OVERRIDES,
-    tokenAdjacencyGraph: DEFAULT_TOKEN_ADJACENCY_GRAPH_BY_CHAIN_ID[CHAIN_ID],
 };
 
 export const defaultHttpServiceConfig: HttpServiceConfig = {
