+++ test/swap_test.ts
@@ -14,7 +14,7 @@ import supertest from 'supertest';
 import { getAppAsync } from '../src/app';
 import { getDefaultAppDependenciesAsync } from '../src/runners/utils';
 import { AppDependencies } from '../src/types';
-import { BUY_SOURCE_FILTER_BY_CHAIN_ID, ChainId, ERC20BridgeSource, LimitOrderFields } from '../src/asset-swapper';
+import { BUY_SOURCE_FILTER_BY_CHAIN_ID, ERC20BridgeSource, LimitOrderFields } from '../src/asset-swapper';
 import * as config from '../src/config';
 import { AFFILIATE_FEE_TRANSFORMER_GAS, GAS_LIMIT_BUFFER_MULTIPLIER, SWAP_PATH } from '../src/constants';
 import { getDBConnectionOrThrow } from '../src/db_connection';
@@ -41,6 +41,7 @@ import { constructRoute, httpGetAsync } from './utils/http_utils';
 import { MockOrderWatcher } from './utils/mock_order_watcher';
 import { getRandomSignedLimitOrderAsync } from './utils/orders';
 import { StatusCodes } from 'http-status-codes';
+import { ChainId } from '@0x/contract-addresses';
 
 // Force reload of the app avoid variables being polluted between test suites
 // Warning: You probably don't want to move this
